Document the initd-forever setup correctly
We suggest using the `-e` flag to set environment variables, but this flag
only lets us set the NODE_ENV variable.  We do need to set
`SIDE_BY_SIDE_HOME` in the init.d script, but to do so we have to do it
manually for now.

Also we change the URL for editing mappings to point at
`publishing.service.gov.uk` instead of `alphagov.co.uk`.