Update xip.io to nip.io
The xip.io homepage is currently blocked at GDS. You can, however still
use it as per the docs to view the side by side browser when doing
development. Nip.io seems to be a drop in replacement that isn't
blocked.

Update the documentation so that developers who aren't familiar with
xip.io don't get roadblocked when they curiously click on the link in
the documentation.